refactor name package, refactor activity and member
